requirements.txt
google_generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community
fastapi
uvicorn
python-multipart